Disallow registration when the proposed email is half-retired
Our learner retirement implementation shall allow re-use of email
addresses, but we currently do not disallow re-use of emails for
learners whose retirement is still in-progress (i.e. their retirement
state is between PENDING and LMS_COMPLETE inclusive).

The time between a user initiating retirement, and the jenkins job
actually picking up the user and driving their account retirement might
be as long as 1 hour, so this is a serious concern.

Addresses EDUCATOR-2824.